pyproject.toml
[project]
name = "iacs-sentiometer-study"
version = "0.1.0"
description = "IACS Sentiometer validation study: LSL streaming, experimental tasks, and data collection"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Nicco Reggente", email = "nicco@iacscience.org" },
]

dependencies = [
    "pyserial>=3.5",
    "pylsl>=1.16.2",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
tasks = [
    "psychopy>=2024.1",
    "numpy>=1.26",
    "pandas>=2.2",
]

[project.scripts]
sentiometer = "sentiometer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sentiometer", "src/tasks"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
